requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.6
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.0.4
django-cors-headers==3.12.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
dlib==19.24.0
drf-yasg==1.20.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.1.0
kombu==5.2.4
MarkupSafe==2.1.1
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.1
prompt-toolkit==3.0.29
pyparsing==3.0.9
pytz==2022.1
redis==4.3.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
torch==1.11.0
torchvision==0.12.0
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
pandas